Move all the graphics related code to TournamentGame
